+++ photutils/aperture/ellipse.py
@@ -132,14 +132,15 @@ class EllipticalAperture(EllipticalMaskMixin, PixelAperture):
 
     Parameters
     ----------
-    positions : array_like or `~astropy.units.Quantity`
+    positions : array_like
         The pixel coordinates of the aperture center(s) in one of the
         following formats:
 
             * single ``(x, y)`` pair as a tuple, list, or `~numpy.ndarray`
             * tuple, list, or `~numpy.ndarray` of ``(x, y)`` pairs
             * `~astropy.units.Quantity` instance of ``(x, y)`` pairs in
-              pixel units
+              pixel units (this is Deprecated and will be removed in a
+              future version)
 
     a : float
         The semimajor axis of the ellipse in pixels.
@@ -263,14 +264,15 @@ class EllipticalAnnulus(EllipticalMaskMixin, PixelAperture):
 
     Parameters
     ----------
-    positions : array_like or `~astropy.units.Quantity`
+    positions : array_like
         The pixel coordinates of the aperture center(s) in one of the
         following formats:
 
             * single ``(x, y)`` pair as a tuple, list, or `~numpy.ndarray`
             * tuple, list, or `~numpy.ndarray` of ``(x, y)`` pairs
             * `~astropy.units.Quantity` instance of ``(x, y)`` pairs in
-              pixel units
+              pixel units (this is Deprecated and will be removed in a
+              future version)
 
     a_in : float
         The inner semimajor axis of the elliptical annulus in pixels.
@@ -432,12 +434,12 @@ class SkyEllipticalAperture(SkyAperture):
         either scalar coordinates or an array of coordinates.
 
     a : scalar `~astropy.units.Quantity`
-        The semimajor axis of the ellipse, either in angular or pixel
-        units.
+        The semimajor axis of the ellipse in angular units. Pixel units
+        are now deprecated.
 
     b : scalar `~astropy.units.Quantity`
-        The semiminor axis of the ellipse, either in angular or pixel
-        units.
+        The semiminor axis of the ellipse in angular units. Pixel units
+        are now deprecated.
 
     theta : scalar `~astropy.units.Quantity`, optional
         The position angle (in angular units) of the ellipse semimajor
@@ -455,8 +457,8 @@ class SkyEllipticalAperture(SkyAperture):
 
     _params = ('positions', 'a', 'b', 'theta')
     positions = SkyCoordPositions('The center position(s) in sky coordinates.')
-    a = ScalarAngleOrPixel('The semimajor axis, in angular or pixel units.')
-    b = ScalarAngleOrPixel('The semiminor axis, in angular or pixel units.')
+    a = ScalarAngleOrPixel('The semimajor axis in angular units.')
+    b = ScalarAngleOrPixel('The semiminor axis in angular units.')
     theta = ScalarAngle('The position angle in angular units of the ellipse '
                         'semimajor axis.')
 
@@ -505,17 +507,21 @@ class SkyEllipticalAnnulus(SkyAperture):
         either scalar coordinates or an array of coordinates.
 
     a_in : scalar `~astropy.units.Quantity`
-        The inner semimajor axis, either in angular or pixel units.
+        The inner semimajor axis in angular units. Pixel units are now
+        deprecated.
 
     a_out : scalar `~astropy.units.Quantity`
-        The outer semimajor axis, either in angular or pixel units.
+        The outer semimajor axis in angular units. Pixel units are now
+        deprecated.
 
     b_out : scalar `~astropy.units.Quantity`
-        The outer semiminor axis, either in angular or pixel units.
+        The outer semiminor axis in angular units. Pixel units are now
+        deprecated.
 
     b_in : `None` or scalar `~astropy.units.Quantity`
-        The inner semiminor axis, either in angular or pixel units.
-        If `None`, then the inner semiminor axis is calculated as:
+        The inner semiminor axis in angular units. Pixel units are
+        now deprecated. If `None`, then the inner semiminor axis is
+        calculated as:
 
             .. math:: b_{in} = b_{out}
                 \left(\frac{a_{in}}{a_{out}}\right)
@@ -537,14 +543,10 @@ class SkyEllipticalAnnulus(SkyAperture):
 
     _params = ('positions', 'a_in', 'a_out', 'b_in', 'b_out', 'theta')
     positions = SkyCoordPositions('The center position(s) in sky coordinates.')
-    a_in = ScalarAngleOrPixel('The inner semimajor axis, in angular or pixel '
-                              'units.')
-    a_out = ScalarAngleOrPixel('The outer semimajor axis, in angular or pixel '
-                               'units.')
-    b_in = ScalarAngleOrPixel('The inner semiminor axis, in angular or pixel '
-                              'units.')
-    b_out = ScalarAngleOrPixel('The outer semiminor axis, in angular or pixel '
-                               'units.')
+    a_in = ScalarAngleOrPixel('The inner semimajor axis in angular units.')
+    a_out = ScalarAngleOrPixel('The outer semimajor axis in angular units.')
+    b_in = ScalarAngleOrPixel('The inner semiminor axis in angular units.')
+    b_out = ScalarAngleOrPixel('The outer semiminor axis in angular units.')
     theta = ScalarAngle('The position angle in angular units of the ellipse '
                         'semimajor axis.')